Consolidated Param and Prior classes into the model file
